FileOpener: a bit fix.

+++ tool/wlog/util.hpp
@@ -253,7 +253,8 @@ public:
 };
 
 /**
- * A simple file writer.
+ * A simple file opener.
+ * close() will be called in destructor when you forget to call it.
  */
 class FileOpener
 {
@@ -1013,7 +1014,7 @@ template <typename ByteType>
 void printByteArray(ByteType *data, size_t size)
 {
     for (size_t i = 0; i < size; i++) {
-        ::printf("%02x", data[i]);
+        ::printf("%02x", static_cast<uint8_t>(data[i]));
 
         if (i % 64 == 63) { ::printf("\n"); }
     }
